+++ src/SecucardConnect/SecucardConnect.php
@@ -33,7 +33,7 @@ final class SecucardConnect
     /**
      * SDK version
      */
-    const VERSION = '1.25.0';
+    const VERSION = '1.26.0';
 
     /**
      * @var OAuthProvider
@@ -109,7 +109,6 @@ final class SecucardConnect
         $this->logger = $logger == null ? new Logger(null, false) : $logger;
 
         $config->isValid();
-        $this->logger->debug('Using config: ' . json_encode($config->toArray(), JSON_PRETTY_PRINT));
         $this->config = $config;
 
         // Create the default common storage if necessary